UI Changes for Location Search
